move python 3 tests to separate file
